+++ src/main/java/zone/moddev/mc/ironagefurniture/api/entity/ThrownLavaLamp.java
@@ -12,7 +12,9 @@ import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.projectile.ProjectileItemEntity;
 import net.minecraft.item.Item;
+import net.minecraft.network.IPacket;
 import net.minecraft.world.World;
+import net.minecraftforge.fml.network.NetworkHooks;
 import net.minecraft.block.Blocks;
 import net.minecraft.block.BlockState;
 import net.minecraft.state.properties.BlockStateProperties;
@@ -36,6 +38,11 @@ public class ThrownLavaLamp extends ProjectileItemEntity {
     @Override
     protected Item func_213885_i() {
         return BlockObjectHolder.light_metal_ironage_block_floor_lava_clear.asItem();
+    }
+
+    @Override
+    public IPacket<?> createSpawnPacket() {
+        return NetworkHooks.getEntitySpawningPacket(this);
     }
 
     @Override